% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmarkcorr.lpp.R
\name{fmarkcorr.lpp}
\alias{fmarkcorr.lpp}
\title{Functional mark correlation function for point patterns over a linear network}
\usage{
fmarkcorr.lpp(
  X,
  r = NULL,
  normalise = FALSE,
  f = function(m1, m2) {
     m1 * m2
 },
  ftype = c("corr", "vario", "rcorr", "schlather", "Beisbart"),
  method = c("density", "loess"),
  ...
)
}
\arguments{
\item{X}{an object of class lpp}

\item{r}{Optional. Numeric vector. The values of the argument r at which the mark correlation function should be evaluated.}

\item{normalise}{If normalise=FALSE, compute only the numerator of the expression for the mark correlation.}

\item{f}{Optional. Test function f used in the definition of the mark correlation function. An R function with at least two arguments. There is a sensible default.}

\item{ftype}{type of test function used in argument f. Currently any selection of the options "corr","vario","rcorr","schlather","equ","Beisbart"}

\item{method}{type of smoothing, either density or loess.}
}
\value{
a data.frame which shows mark correlation functions evaluated at each time as well as the overall values.
}
\description{
Functional mark correlation function for point patterns over a linear network
}
\examples{
L <- spiders$domain
X <- runiflpp(150,L=L)
m <- t(replicate(150,runif(513)))
marks(X) <- as.data.frame(m)
Fcor <- fmarkcorr.linnet(X,r, ftype = "corr", method = "density" ,normalise = TRUE)
plot(Fcor$r,Fcor$gw,type = "l")
}
\references{
Eckardt, M., Mateu, J., and Moradi, M. (2023) Point processes on linear networks with function-valued marks.
}
\author{
Mehdi Moradi \email{m2.moradi@yahoo.com} and Matthias Eckardt
}
