% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearinhommarkk.lpp.R
\name{linearinhommarkk.lpp}
\alias{linearinhommarkk.lpp}
\title{Mark-Weighted inhomogeneous K Function for point patterns over a linear network}
\usage{
linearinhommarkk.lpp(X, r = r, lambda = lambda, normalize = FALSE, ...)
}
\arguments{
\item{X}{an object of class lpp}

\item{r}{Optional. Numeric vector. The values of the argument r at which the mark correlation function should be evaluated.}

\item{lambda}{Intensity values at data points.}

\item{normalize}{Logical.}
}
\value{
a numeric vector.
}
\description{
Mark-Weighted inhomogeneous K Function for point patterns over a linear network
}
\examples{
 X <- rpoislpp(10,simplenet)
 r <- seq(0,boundingradius(simplenet),length.out=513)
 dx <- densityQuick.lpp(X,at = "points")
 linearinhommarkk.lpp(X,r=r,lambda=dx)
}
\references{
Eckardt, M., and Moradi, M. (2023) Marked point processes on linear networks.
}
\author{
Mehdi Moradi \email{m2.moradi@yahoo.com} and Matthias Eckardt
}
