% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearmarkk.lpp.R
\name{linearmarkk.lpp}
\alias{linearmarkk.lpp}
\title{Mark-Weighted homogeneous K Function for point patterns over a linear network}
\usage{
linearmarkk.lpp(X, r = r)
}
\arguments{
\item{X}{an object of class lpp}

\item{r}{Optional. Numeric vector. The values of the argument r at which the mark correlation function should be evaluated.}
}
\value{
a numeric vector.
}
\description{
Mark-Weighted homogeneous K Function for point patterns over a linear network
}
\examples{
 X <- rpoislpp(10,simplenet)
 r <- seq(0,boundingradius(simplenet),length.out=513)
 linearmarkk.lpp(X,r=r)
}
\references{
Eckardt, M., and Moradi, M. (2023) Marked point processes on linear networks.
}
\author{
Mehdi Moradi \email{m2.moradi@yahoo.com} and Matthias Eckardt
}
