% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markcorr.lpp.R
\name{markcorr.lpp}
\alias{markcorr.lpp}
\title{Mark correlation function for point patterns over a linear network}
\usage{
markcorr.lpp(
  X,
  r = NULL,
  normalise = TRUE,
  f = function(m1, m2) {
     m1 * m2
 },
  ftype = c("corr", "vario", "rcorr", "schlather", "equ", "Beisbart"),
  method = c("density", "loess"),
  ...
)
}
\arguments{
\item{X}{an object of class lpp}

\item{r}{Optional. Numeric vector. The values of the argument r at which the mark correlation function should be evaluated.}

\item{normalise}{If normalise=FALSE, compute only the numerator of the expression for the mark correlation.}

\item{f}{Optional. Test function f used in the definition of the mark correlation function. An R function with at least two arguments. There is a sensible default.}

\item{ftype}{type of test function used in argument f. Currently any selection of the options "corr","vario","rcorr","schlather","equ","Beisbart"}

\item{method}{type of smoothing, either density or loess.}
}
\value{
a data.frame which gives the empirical mark correlation function and the distance vector r where the mark correlation finction is evaluated.
}
\description{
Mark correlation function for point patterns over a linear network
}
\examples{
 X <- rpoislpp(10,simplenet)
 marks(X) <- runif(npoints(X),10,11)
 markcorr.lpp(X,r=r,ftype = "corr",f=function(m1,m2){m1*m2})
}
\references{
Eckardt, M., and Moradi, M. (2024) Marked spatial point processes: current state and extensions to point processes on linear networks
}
\author{
Mehdi Moradi \email{m2.moradi@yahoo.com} and Matthias Eckardt
}
